% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{.enrichment}
\alias{.enrichment}
\title{Generic enrichment function}
\usage{
.enrichment(n.outliers.w.var, n.outliers, n.non.outliers.w.var,
  n.non.outliers)
}
\description{
Calculates relative risk given the number of outliers with a variant, the number of outliers, the number
of non-outliers with a variant, and the number of non outliers.
}
\details{
It is used in a bunch of different things and is internal
}
