% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{.unique_IG_pairs}
\alias{.unique_IG_pairs}
\title{Unique individual-gene pairs in a data-frame}
\usage{
.unique_IG_pairs(x)
}
\description{
Internal function that returns the number of unique individual-gene pairs from a data-frame
}
